+++ src/commonMain/kotlin/at/posselt/pfrpg2e/data/kingdom/activities/ActivityDcType.kt
@@ -10,6 +10,8 @@ enum class ActivityDcType: ValueEnum, Translatable {
     CONTROL,
     CUSTOM,
     NONE,
+    NEGOTIATION,
+    NEGOTIATION_OR_CONTROL,
     SCOUTING;
 
     companion object {