+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/dialogs/ArmyTacticsBrowser.kt
@@ -127,6 +127,7 @@ private class ArmyTacticsBrowser(
                 ),
                 overrideDc = getLevelBasedDC(item.level),
                 selectedLeader = game.getActiveLeader(),
+                groups = emptyArray(),
             )
         }
     }